# lib/httpServer/index.js
const express = require('express');
const bodyParser = require('body-parser');
const swaggerUi = require('swagger-ui-express');
const R = require('ramda');
const path = require('path');
const swaggerDocument = require('./swagger.json');
const Block = require('../blockchain/block');
const Transaction = require('../blockchain/transaction');
const TransactionAssertionError = require('../blockchain/transactionAssertionError');
const BlockAssertionError = require('../blockchain/blockAssertionError');
const HTTPError = require('./httpError');
const ArgumentError = require('../util/argumentError');
const CryptoUtil = require('../util/cryptoUtil');
const timeago = require('timeago.js');

class HttpServer {
    constructor(node, blockchain, operator, miner) {
        this.app = express();//Swagger UI Express!

        const projectWallet = (wallet) => {
            return {
                id: wallet.id,
                addresses: R.map((keyPair) => {
                    return keyPair.publicKey;
                }, wallet.keyPairs)
            };
        };

        this.app.use(bodyParser.urlencoded());
        this.app.use(bodyParser.json());

        this.app.set('view engine', 'pug');
        this.app.set('views', path.join(__dirname, 'views'));
        this.app.locals.formatters = {
            time: (rawTime) => {
                const timeInMS = new Date(rawTime * 1000);
                return `${timeInMS.toLocaleString()} - ${timeago().format(timeInMS)}`;
            },
            hash: (hashString) => {
                return hashString != '0' ? `${hashString.substr(0, 5)}...${hashString.substr(hashString.length - 5, 5)}` : '<empty>';
            },
            amount: (amount) => amount.toLocaleString()
        };
        this.app.use('/api-docs', swaggerUi.serve, swaggerUi.setup(swaggerDocument));

        this.app.get('/blockchain', (req, res) => {
            if (req.headers['accept'] && req.headers['accept'].includes('text/html'))
                res.render('blockchain/index.pug', {
                    pageTitle: 'Blockchain',
                    blocks: blockchain.getAllBlocks()
                });
            else
                throw new HTTPError(400, 'Accept content not supported');
        });

        this.app.get('/rydecoin', (req, res) => {
            if (req.headers['accept'] && req.headers['accept'].includes('text/html'))
                res.render('rydecoin/current_rides/index.pug', {
                    pageTitle: 'Current Rides',
                    blocks: blockchain.getAllBlocks()
                });
            else
                throw new HTTPError(400, 'Accept content not supported');
        });

        this.app.get('/rydecoin/create_account', (req, res) => {
            let managerAddress = operator.getManagerAddress();
            if (req.headers['accept'] && req.headers['accept'].includes('text/html'))
                res.render('rydecoin/create_account/index.pug', {
                    pageTitle: 'Create Account',
                    message: '',
                    walletId: '',
                    address: ''
                });
            else
                throw new HTTPError(400, 'Accept content not supported');
        });

        this.app.post('/rydecoin/create_account', (req, res) => {
            let message = '';
            let walletId = '';
            let newAddress = '';
            let password = req.body.password;
            let confirmPassword = req.body.confirmPassword;

            if (R.match(/\w+/g, password).length <= 4) {
                message = 'ERROR: Password must contain more than 4 words';
            } else if (password != confirmPassword) {
                message = 'ERROR: Passwords do not match'
            } else {
                try {
                    walletId = operator.createWalletFromPassword(password).id
                    newAddress = operator.generateAddressForWallet(walletId);
                    let managerWalletId = operator.getManagerWalletId();
                    let managerAddress = operator.getManagerAddress();
                    if (blockchain.getUnspentTransactionsForAddress(managerAddress) < 1000) {
                        miner.mine(managerAddress, managerAddress)
                             .then((newBlock) => {
                                 newBlock = Block.fromJson(newBlock);
                                 blockchain.addBlock(newBlock);
                             });
                    }

                    let newTransaction = operator.createTransaction(managerWalletId, managerAddress, newAddress, 1000, managerAddress);

                    newTransaction.check();

                    let transactionCreated = blockchain.addTransaction(Transaction.fromJson(newTransaction));
                    message = 'Successfully created account! Please save your wallet_id and address'
                } catch (ex) {
                    console.warn(ex.message);
                    message = 'Successfully created account! Please save your wallet_id and address. ERROR: Could not transfer money from manager.';
                }
            }
            res.render('rydecoin/create_account/index.pug', {
                pageTitle: 'Create Account',
                message: message,
                walletId: walletId,
                address: newAddress
            })
        });

        this.app.get('/rydecoin/post_ride', (req, res) => {
            if (req.headers['accept'] && req.headers['accept'].includes('text/html'))
                res.render('rydecoin/post_ride/index.pug', {
                    pageTitle: 'Post Ride',
                    message: ''
                });
            else
                throw new HTTPError(400, 'Accept content not supported');
        });

        this.app.post('/rydecoin/post_ride', (req, res) => {
            let message = '';

            let walletId = req.body.walletId;
            let password = req.body.password;
            let address = req.body.address;
            let dateOfJourney = req.body.dateOfJourney;
            let from = req.body.from;
            let to = req.body.to;
            let space = req.body.space;
            let rideCost = req.body.rideCost;

            let rideInfo =  {
              "space": space,
              "dateOfJourney": dateOfJourney,
              "from": from,
              "to": to,
              "rideCost": rideCost
            }
            let passwordHash = CryptoUtil.hash(password);
            let amount = space*rideCost;
            let managerAddress = operator.getManagerAddress();

            try {
                if (!operator.checkWalletPassword(walletId, passwordHash)) throw new HTTPError(403, `Invalid password for wallet '${walletId}'`);

                let newTransaction = operator.createTransaction_v1(walletId, address, managerAddress, amount, address, rideInfo, 'create_ride');

                newTransaction.check();

                let transactionCreated = blockchain.addTransaction(Transaction.fromJson(newTransaction));
                message = "Ride successfully posted!"
            } catch (ex) {
                message = `ERROR: ${ex.message}`
            }
            res.render('rydecoin/post_ride/index.pug', {
                pageTitle: 'Post Ride',
                message: message
            });
        });


        this.app.post('/rydecoin/book_ride/',(req,res) => {
            
            let walletId = req.body.walletId;
            let password = req.body.password;
            let addressId = req.body.addressId;
            let rideId = req.body.rideId;
            let amount = req.body.amount;
            //console.info(`Booking the ride :`+walletId+'....'+addressId+'....'+rideId+'..'+amount);
            let walletAddresses = operator.getAddressesForWallet(walletId);
            let passwordHash = CryptoUtil.hash(password);

            let rideInfo =  {
                "rideId":rideId
            }

            if (walletAddresses.indexOf(req.body.addressId) === -1)  throw new HTTPError(403, `Invalid address for wallet '${walletId}'`);

            if (password == null) throw new HTTPError(401, 'Wallet\'s password is missing.');

            try{
                let managerAddress = operator.getManagerAddress();

                if (!operator.checkWalletPassword(walletId, passwordHash)) throw new HTTPError(403, `Invalid password for wallet '${walletId}'`);

                let newTransaction =operator.createTransaction_v1(walletId, addressId , managerAddress, amount , addressId , rideInfo , 'book_ride');

                newTransaction.check();
                
                let transactionCreated = blockchain.addTransaction(Transaction.fromJson(newTransaction));
                
                console.log('Ride booked');
                


            }catch(e){
                console.warn(e.message);
            }
            //res.redirect('/rydecoin');
            //res.status(200).send('Ride Booked');
            res.redirect(200, '/rydecoin');

        });


        this.app.get('/blockchain/blocks', (req, res) => {
            res.status(200).send(blockchain.getAllBlocks());
        });

        this.app.get('/blockchain/appState', (req, res) => {
            res.status(200).send(blockchain.getAppState());
        });

        this.app.get('/blockchain/blocks/latest', (req, res) => {
            let lastBlock = blockchain.getLastBlock();
            if (lastBlock == null) throw new HTTPError(404, 'Last block not found');

            res.status(200).send(lastBlock);
        });

        this.app.put('/blockchain/blocks/latest', (req, res) => {
            let requestBlock = Block.fromJson(req.body);
            let result = node.checkReceivedBlock(requestBlock);

            if (result == null) res.status(200).send('Requesting the blockchain to check.');
            else if (result) res.status(200).send(requestBlock);
            else throw new HTTPError(409, 'Blockchain is update.');
        });

        this.app.get('/blockchain/blocks/:hash([a-zA-Z0-9]{64})', (req, res) => {
            let blockFound = blockchain.getBlockByHash(req.params.hash);
            if (blockFound == null) throw new HTTPError(404, `Block not found with hash '${req.params.hash}'`);

            res.status(200).send(blockFound);
        });

        this.app.get('/blockchain/blocks/:index', (req, res) => {
            let blockFound = blockchain.getBlockByIndex(parseInt(req.params.index));
            if (blockFound == null) throw new HTTPError(404, `Block not found with index '${req.params.index}'`);

            res.status(200).send(blockFound);
        });

        this.app.get('/blockchain/blocks/transactions/:transactionId([a-zA-Z0-9]{64})', (req, res) => {
            let transactionFromBlock = blockchain.getTransactionFromBlocks(req.params.transactionId);
            if (transactionFromBlock == null) throw new HTTPError(404, `Transaction '${req.params.transactionId}' not found in any block`);

            res.status(200).send(transactionFromBlock);
        });

        this.app.get('/blockchain/transactions', (req, res) => {
            if (req.headers['accept'] && req.headers['accept'].includes('text/html'))
                res.render('blockchain/transactions/index.pug', {
                    pageTitle: 'Unconfirmed Transactions',
                    transactions: blockchain.getAllTransactions()
                });
            else
                res.status(200).send(blockchain.getAllTransactions());
        });

        this.app.post('/blockchain/transactions', (req, res) => {
            let requestTransaction = Transaction.fromJson(req.body);
            let transactionFound = blockchain.getTransactionById(requestTransaction.id);

            if (transactionFound != null) throw new HTTPError(409, `Transaction '${requestTransaction.id}' already exists`);

            try {
                let newTransaction = blockchain.addTransaction(requestTransaction);
                res.status(201).send(newTransaction);
            } catch (ex) {
                if (ex instanceof TransactionAssertionError) throw new HTTPError(400, ex.message, requestTransaction, ex);
                else throw ex;
            }
        });

        this.app.get('/blockchain/transactions/unspent', (req, res) => {
            res.status(200).send(blockchain.getUnspentTransactionsForAddress(req.query.address));
        });

        this.app.get('/operator/wallets', (req, res) => {
            let wallets = operator.getWallets();

            let projectedWallets = R.map(projectWallet, wallets);

            res.status(200).send(projectedWallets);
        });

        this.app.post('/operator/wallets', (req, res) => {
            let password = req.headers.password;
            if (R.match(/\w+/g, password).length <= 4) throw new HTTPError(400, 'Password must contain more than 4 words');

            let newWallet = operator.createWalletFromPassword(password);

            let projectedWallet = projectWallet(newWallet);

            res.status(201).send(projectedWallet);
        });

        this.app.get('/operator/wallets/:walletId', (req, res) => {
            let walletFound = operator.getWalletById(req.params.walletId);
            if (walletFound == null) throw new HTTPError(404, `Wallet not found with id '${req.params.walletId}'`);

            let projectedWallet = projectWallet(walletFound);

            res.status(200).send(projectedWallet);
        });

        this.app.post('/operator/wallets/:walletId/transactions', (req, res) => {
            let walletId = req.params.walletId;
            let password = req.headers.password;

            if (password == null) throw new HTTPError(401, 'Wallet\'s password is missing.');
            let passwordHash = CryptoUtil.hash(password);

            try {
                if (!operator.checkWalletPassword(walletId, passwordHash)) throw new HTTPError(403, `Invalid password for wallet '${walletId}'`);

                let newTransaction = operator.createTransaction(walletId, req.body.fromAddress, req.body.toAddress, req.body.amount, req.body['changeAddress'] || req.body.fromAddress);

                newTransaction.check();

                let transactionCreated = blockchain.addTransaction(Transaction.fromJson(newTransaction));
                res.status(201).send(transactionCreated);
            } catch (ex) {
                if (ex instanceof ArgumentError || ex instanceof TransactionAssertionError) throw new HTTPError(400, ex.message, walletId, ex);
                else throw ex;
            }
        });

        this.app.get('/operator/wallets/:walletId/addresses', (req, res) => {
            let walletId = req.params.walletId;
            try {
                let addresses = operator.getAddressesForWallet(walletId);
                res.status(200).send(addresses);
            } catch (ex) {
                if (ex instanceof ArgumentError) throw new HTTPError(400, ex.message, walletId, ex);
                else throw ex;
            }
        });

        this.app.post('/operator/wallets/:walletId/addresses', (req, res) => {
            let walletId = req.params.walletId;
            let password = req.headers.password;

            if (password == null) throw new HTTPError(401, 'Wallet\'s password is missing.');
            let passwordHash = CryptoUtil.hash(password);

            try {
                if (!operator.checkWalletPassword(walletId, passwordHash)) throw new HTTPError(403, `Invalid password for wallet '${walletId}'`);

                let newAddress = operator.generateAddressForWallet(walletId);
                res.status(201).send({ address: newAddress });
            } catch (ex) {
                if (ex instanceof ArgumentError) throw new HTTPError(400, ex.message, walletId, ex);
                else throw ex;
            }
        });

        this.app.get('/operator/:addressId/balance', (req, res) => {
            let addressId = req.params.addressId;

            try {
                let balance = operator.getBalanceForAddress(addressId);
                res.status(200).send({ balance: balance });
            } catch (ex) {
                if (ex instanceof ArgumentError) throw new HTTPError(404, ex.message, { addressId }, ex);
                else throw ex;
            }
        });


        this.app.post('/operator/wallets/:walletId/postRide', (req, res) => {
            let walletId = req.params.walletId;
            let password = req.headers.password;

            if (password == null) throw new HTTPError(401, 'Wallet\'s password is missing.');
            let passwordHash = CryptoUtil.hash(password);

            try {
                if (!operator.checkWalletPassword(walletId, passwordHash)) throw new HTTPError(403, `Invalid password for wallet '${walletId}'`);

                let newTransaction = operator.createTransaction_v1(walletId, req.body.fromAddress, req.body.toAddress, 
                    req.body.amount, req.body['changeAddress'] || req.body.fromAddress,
                    req.body.rideInfo, 'create_ride');

                newTransaction.check();

                let transactionCreated = blockchain.addTransaction(Transaction.fromJson(newTransaction));
                res.status(201).send(transactionCreated);
            } catch (ex) {
                if (ex instanceof ArgumentError || ex instanceof TransactionAssertionError) throw new HTTPError(400, ex.message, walletId, ex);
                else throw ex;
            }
        });
        //As put is idempotent , preferrable over post!
        this.app.post('/operator/wallets/:walletId/bookRide',(req,res)=>{
            let walletId = req.params.walletId;
            let password = req.headers.password;

            if(password == null) throw new HTTPError(401, 'Wallet\'s password is missing !');
            let passwordHash = CryptoUtil.hash(password);
            
            try{
                if (!operator.checkWalletPassword(walletId, passwordHash)) throw new HTTPError(403, `Invalid password for wallet '${walletId}'`);//'' vs `
                //below puts the transaction on the TransactionJSON!!
                let newBookTransaction =operator.createTransaction_v1(walletId,req.body.fromAddress,req.body.toAddress,req.body.amount,
                    req.body['changeAddress'] || req.body.fromAddress,req.body.rideInfo,'book_ride');


                //Here itself we will check whther the hash is correct of the transaction
                newBookTransaction.check();

                let bookTransactionCreated= blockchain.addTransaction(Transaction.fromJson(newBookTransaction));
                res.status(201).send(bookTransactionCreated);

            }catch(ex){
                if(ex instanceof ArgumentError ||  ex instanceof TransactionAssertionError) throw new HTTPError(400, ex.message, walletId,ex);
                else throw ex;
            }


        });

        this.app.post('/operator/wallets/:walletId/finishRide', (req, res) => {
            let walletId = req.params.walletId;
            let password = req.headers.password;
            let walletAddresses = operator.getAddressesForWallet(walletId);
            if (walletAddresses.indexOf(req.body.userAddress) === -1)
                throw new HTTPError(403, `Invalid address for wallet '${walletId}'`);

            if (password == null) throw new HTTPError(401, 'Wallet\'s password is missing.');
            let passwordHash = CryptoUtil.hash(password);

            try {
                if (!operator.checkWalletPassword(walletId, passwordHash)) throw new HTTPError(403, `Invalid password for wallet '${walletId}'`);

                //Find the relevant trigger transaction and get values

                let amount = blockchain.getRideTransaction(req.body.rideId,req.body.userAddress);
                let fromAddress = operator.getManagerAddress();
                let operatorWalletId = operator.getManagerWalletId();
                let newTransaction = operator.createTransaction_v1(operatorWalletId, fromAddress, req.body.userAddress, 
                    amount, fromAddress, null, 'finish_ride');

                newTransaction.check();

                let transactionCreated = blockchain.addTransaction(Transaction.fromJson(newTransaction));
                res.status(201).send(transactionCreated);
            } catch (ex) {
                if (ex instanceof ArgumentError || ex instanceof TransactionAssertionError) throw new HTTPError(400, ex.message, walletId, ex);
                else throw ex;
            }
        });

        this.app.get('/node/peers', (req, res) => {
            res.status(200).send(node.peers);
        });

        this.app.post('/node/peers', (req, res) => {
            let newPeer = node.connectToPeer(req.body);
            res.status(201).send(newPeer);
        });

        this.app.get('/node/transactions/:transactionId([a-zA-Z0-9]{64})/confirmations', (req, res) => {
            node.getConfirmations(req.params.transactionId)
                .then((confirmations) => {
                    res.status(200).send({ confirmations: confirmations });
                });
        });

        this.app.post('/miner/mine', (req, res, next) => {
            miner.mine(req.body.rewardAddress, req.body['feeAddress'] || req.body.rewardAddress)
                .then((newBlock) => {
                    newBlock = Block.fromJson(newBlock);
                    blockchain.addBlock(newBlock);
                    res.status(201).send(newBlock);
                })
                .catch((ex) => {
                    if (ex instanceof BlockAssertionError && ex.message.includes('Invalid index')) next(new HTTPError(409, 'A new block were added before we were able to mine one'), null, ex);
                    else next(ex);
                });
        });

        this.app.use(function (err, req, res, next) {  // eslint-disable-line no-unused-vars
            if (err instanceof HTTPError) res.status(err.status);
            else res.status(500);
            res.send(err.message + (err.cause ? ' - ' + err.cause.message : ''));
        });
    }

    listen(host, port) {
        return new Promise((resolve, reject) => {
            this.server = this.app.listen(port, host, (err) => {
                if (err) reject(err);
                console.info(`Listening http on port: ${this.server.address().port}, to access the API documentation go to http://${host}:${this.server.address().port}/api-docs/`);
                resolve(this);
            });
        });
    }

    stop() {
        return new Promise((resolve, reject) => {
            this.server.close((err) => {
                if (err) reject(err);
                console.info('Closing http');
                resolve(this);
            });
        });
    }
}

module.exports = HttpServer;